+++ src/league/league.py
@@ -3,6 +3,7 @@ import collections
 
 import numpy as np
 
+from league.payoff import Payoff
 from league.roles.exploiters import MainExploiter, LeagueExploiter
 from league.roles.players import MainPlayer, Player
 
@@ -14,8 +15,8 @@ class Agent(object):
     architecture.
     """
 
-    def __init__(self, race, initial_weights):
-        self.race = race # TODO: replace race with team composition he resides in?
+    def __init__(self, team_plan, initial_weights):
+        self.team_plan = team_plan
         self.steps = 0
         self.weights = initial_weights
 
@@ -47,87 +48,38 @@ class Agent(object):
         return policy_logits, baselines
 
 
-class Payoff:
-
-    def __init__(self):
-        self._players = []
-        self._wins = collections.defaultdict(lambda: 0)
-        self._draws = collections.defaultdict(lambda: 0)
-        self._losses = collections.defaultdict(lambda: 0)
-        self._games = collections.defaultdict(lambda: 0)
-        self._decay = 0.99
-
-    def _win_rate(self, _home, _away):
-        if self._games[_home, _away] == 0:
-            return 0.5
-
-        return (self._wins[_home, _away] +
-                0.5 * self._draws[_home, _away]) / self._games[_home, _away]
-
-    def __getitem__(self, match):
-        home, away = match
-
-        if isinstance(home, Player):
-            home = [home]
-        if isinstance(away, Player):
-            away = [away]
-
-        win_rates = np.array([[self._win_rate(h, a) for a in away] for h in home])
-        if win_rates.shape[0] == 1 or win_rates.shape[1] == 1:
-            win_rates = win_rates.reshape(-1)
-
-        return win_rates
-
-    def update(self, home, away, result):
-        for stats in (self._games, self._wins, self._draws, self._losses):
-            stats[home, away] *= self._decay
-            stats[away, home] *= self._decay
-
-        self._games[home, away] += 1
-        self._games[away, home] += 1
-        if result == "win":
-            self._wins[home, away] += 1
-            self._losses[away, home] += 1
-        elif result == "draw":
-            self._draws[home, away] += 1
-            self._draws[away, home] += 1
-        else:
-            self._wins[away, home] += 1
-            self._losses[home, away] += 1
-
-    def add_player(self, player):
-        self._players.append(player)
-
-    @property
-    def players(self):
-        return self._players
-
-
 class League(object):
 
     def __init__(self,
                  initial_agents,
-                 main_agents=1,
-                 main_exploiters=1,
-                 league_exploiters=2):
+                 main_agents_n=1,
+                 main_exploiters_n=1,
+                 league_exploiters_n=2):
         self._payoff = Payoff()
         self._learning_agents = []
-        # TODO: This will change since we do not consider races but team compositions in terms of agent/unit selection
-        for race in initial_agents:
-            for _ in range(main_agents):
-                main_agent = MainPlayer(race, initial_agents[race], self._payoff)
+        self._main_agents_n = main_agents_n
+        self._main_exploiters_n = main_exploiters_n
+        self._league_exploiters_n = league_exploiters_n
+        # Setup initial learning agents
+        for plan in initial_agents:
+            # TODO: how to build a set of initial agents so that initial_agents[plan] makes sense?
+            for _ in range(self._main_agents_n):
+                main_agent = MainPlayer(team_plan=plan, agent=initial_agents[plan], payoff=self._payoff)
                 self._learning_agents.append(main_agent)
                 self._payoff.add_player(main_agent.checkpoint())
 
-            for _ in range(main_exploiters):
-                self._learning_agents.append(
-                    MainExploiter(race, initial_agents[race], self._payoff))
-            for _ in range(league_exploiters):
-                self._learning_agents.append(
-                    LeagueExploiter(race, initial_agents[race], self._payoff))
+            for _ in range(self._main_exploiters_n):
+                exploiter = MainExploiter(team_plan=plan, agent=initial_agents[plan], payoff=self._payoff)
+                self._learning_agents.append(exploiter)
+            for _ in range(self._league_exploiters_n):
+                league_exploiter = LeagueExploiter(team_plan=plan, agent=initial_agents[plan], payoff=self._payoff)
+                self._learning_agents.append(league_exploiter)
 
         for player in self._learning_agents:
             self._payoff.add_player(player)
+
+    def roles_per_initial_agent(self):
+        return self._main_agents_n + self._main_exploiters_n + self._league_exploiters_n
 
     def update(self, home, away, result):
         return self._payoff.update(home, away, result)
