+++ src/league/roles/exploiters.py
@@ -1,7 +1,15 @@
+from league.league import Agent
+from league.roles.players import Player, MainPlayer, HistoricalPlayer
+import numpy as np
+
+from league.utils.pfsp import prioritized_fictitious_self_play
+
+
 class MainExploiter(Player):
 
-    def __init__(self, race, agent, payoff):
-        self.agent = Agent(race, agent.get_weights())
+    def __init__(self, team_plan, agent, payoff):
+        super().__init__()
+        self.agent = Agent(team_plan, agent.get_weights())
         self._initial_weights = agent.get_weights()
         self._payoff = payoff
         self._race = agent.race
@@ -19,7 +27,7 @@ class MainExploiter(Player):
 
         historical = [
             player for player in self._payoff.players
-            if isinstance(player, Historical) and player.parent == opponent
+            if isinstance(player, HistoricalPlayer) and player.parent == opponent
         ]
         win_rates = self._payoff[self, historical]
 
@@ -46,8 +54,9 @@ class MainExploiter(Player):
 
 class LeagueExploiter(Player):
 
-    def __init__(self, race, agent, payoff):
-        self.agent = Agent(race, agent.get_weights())
+    def __init__(self, team_plan, agent, payoff):
+        super().__init__()
+        self.agent = Agent(team_plan, agent.get_weights())
         self._initial_weights = agent.get_weights()
         self._payoff = payoff
         self._race = agent.race
@@ -56,7 +65,7 @@ class LeagueExploiter(Player):
     def get_match(self):
         historical = [
             player for player in self._payoff.players
-            if isinstance(player, Historical)
+            if isinstance(player, HistoricalPlayer)
         ]
         win_rates = self._payoff[self, historical]
         return np.random.choice(
@@ -69,12 +78,12 @@ class LeagueExploiter(Player):
         return self._create_checkpoint()
 
     def ready_to_checkpoint(self):
-        steps_passed = self._agent.get_steps() - self._checkpoint_step
+        steps_passed = self.agent.get_steps() - self._checkpoint_step
         if steps_passed < 2e9:
             return False
         historical = [
             player for player in self._payoff.players
-            if isinstance(player, Historical)
+            if isinstance(player, HistoricalPlayer)
         ]
         win_rates = self._payoff[self, historical]
         return win_rates.min() > 0.7 or steps_passed > 4e9
