+++ src/runners/league_runner.py
@@ -187,8 +187,8 @@ class LeagueRunner:
             #
             self.env.render()
 
-            home_reward = reward[0]  # TODO: Assumes global reward and team size of 2 with 2 teams
-            opponent_reward = reward[2]
+            home_reward = reward[0]  # TODO: Assumes global reward and team size of 2
+            opponent_reward = reward[1]
             home_episode_return += home_reward
             opp_episode_return += opponent_reward
